feat: separate phone verification page resources
